Add 01.09.02_tryhtml_styles_background-color.html
##HTML Styling
Every HTML element has a default style (background color is white and
text color is black).

Changing the default style of an HTML element, can be done with the
style attribute.

This example changes the default background color from white to
lightgrey:

The bgcolor attribute, supported in older versions of HTML, is not valid
in HTML5.